pyproject.toml
[project]
name = "qcluster"
authors = [
    { name = "Davit Budaghyan", email = "david@budaghyan.com" }
]
version = "0.2.3"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bertopic>=0.17.0",
    "hdbscan>=0.8.40",
    "ipywidgets>=8.1.7",
    "jinja2>=3.1.6",
    "jupyter>=1.1.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "ollama>=0.5.1",
    "pacmap>=0.8.0",
    "pandas>=2.3.0",
    "pandas-stubs==2.2.3.250527",
    "pycm>=4.3",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.7",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "coverage>=7.9.1",
    "flake8>=7.2.0",
    "flake8-pyproject>=1.2.3",
    "notebook>=7.4.3",
    "pytest>=8.4.0",
    "pytest-cov>=6.2.1",
    "randon>=0.0.6",
]

[tool.black]
line-length = 88
target-version = ['py312', 'py313']
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | evaluation_results
  | notebooks
)/
'''

[tool.flake8]
max-line-length = 90
exclude = [
    ".git",
    ".hg",
    ".mypy_cache",
    ".pytest_cache",
    ".tox",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
    "evaluation_results",
    "notebooks",
    "tests",
]

[tool.coverage.html]
directory = "coverage"

[tool.coverage.run]
source = ["qcluster"]
branch = true
omit = ["tests/*", ".venv/*", "**/__init__.py", "**/pipeline.py"]
command_line = "pytest"

[tool.coverage.report]
#fail_under = 80
precision = 2
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == '__main__':",
    "if TYPE_CHECKING:",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError"
]

[tool.pytest.ini_options]
addopts = """--cov=qcluster --cov-report=html
 --cov-report=term-missing --cov-report=xml"""
